Add files via upload
Python program that interacts with the user to input a child's age. Based on the age provided, the program informs the user of the child's category. Children aged 6 to 7 fall into the 'Poussin' category, those aged 8 to 9 belong to the 'Pupille' category, children aged 10 to 11 are categorized as 'Minime,' and those aged 12 and older are in the 'Cadet' category.