Add files via upload
Python program that asks the user to input a month represented by a number. The program then uses a match statement to determine the number of days in that month. If the month number corresponds to a month with 31 days, it informs the user that the month has 31 days. If it's a month with 30 days, it informs the user accordingly. For February (month 2), it mentions that it has 28 or 29 days depending on leap years. In case of an invalid month number, the program lets the user know that the input is not valid.